Make binjnes work on Windows
* Add new BINJNES_{CLANG,GCC,MSVC} preprocessor defines
* Remove Linux-only compile flags
* Remove Linux-only linked libraries
* Add platform-specific atomic.h (stdatomic for C is not supported in my
  version of MSVC)
* Fix some MSVC compiler warnings
* Repurpose UNREACHABLE as __builtin_unreachable/__assume(0)
* MSVC doesn't allow you to predeclare an indeterminately sized array
  :-/
* Extend vec library to support all uses needed for emulator.c (e.g.
  don't use vector extensions)
* Fix alignment bug in apu_tick. Previously this code was forcing
  unaligned accesses to make it work. This didn't seem to work on MSVC.
  Now it is always aligned and masks off parts of the calculation that
  are not used
* Remove uses of case range extension

+++ src/cartdb.c
@@ -71,7 +71,7 @@ static u32 get_crc(u8 *buf, int len) {
 const CartDbInfo* cartdb_info_from_file(const FileData* file_data) {
   const size_t kHeaderSize = 16;
   const size_t kCartDbLength = ARRAY_SIZE(s_cart_db);
-  u32 crc = get_crc(file_data->data + kHeaderSize, file_data->size - kHeaderSize);
+  u32 crc = get_crc(file_data->data + kHeaderSize, (int)(file_data->size - kHeaderSize));
   const CartDbInfo* begin = &s_cart_db[0];
   const CartDbInfo* end = &s_cart_db[kCartDbLength];
   LOWER_BOUND(const CartDbInfo, found, begin, end, crc, GET_CRC, CMP_LT);
